+++ src/Semirings.jl
@@ -5,11 +5,17 @@ module Semirings
 # Necessary for the logarithmic semiring
 import LogExpFunctions: logaddexp
 
+# Strings extension
+export StringZero
+export ∞
+
 # Semirings
 export Semiring
 export BoolSemiring
 export LogSemiring
 export ProbSemiring
+export ProductSemiring
+export StringSemiring
 export TropicalSemiring
 export UnionConcatSemiring, SymbolSequence
 export val
@@ -23,6 +29,7 @@ export IsOrdered, Ordered, Unordered
 include("semiringtype.jl")
 include("properties.jl")
 include("semiringimpl.jl")
+include("strings.jl")
 
 end